requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
sentence-transformers==2.2.2
torch==2.8.0
faiss-cpu==1.12.0
numpy==2.0.2
pandas==2.2.2
Pillow==11.3.0
